chore: made cypress ci-ready

+++ botfront/cypress/integration/01_initial_setup_dont_change_name/initial_setup.spec.js
@@ -2,7 +2,9 @@
 
 describe('intial setup', function() {
     before(function() {
-        cy.exec('mongo meteor --host localhost:3001 --eval "db.dropDatabase();"');
+        if (!Cypress.env('MODE') || Cypress.env('MODE') !== 'CI_RUN') {
+            cy.exec('mongo meteor --host localhost:3001 --eval "db.dropDatabase();"');
+        }
     });
 
     it('Should create projects when completing the initial setup', () => {
